Minor: code organization
