+++ src/engine/plugins/listener.rs
@@ -4,9 +4,9 @@ use tokio::sync::mpsc::Receiver;
 use crate::prelude::*;
 
 /// Resources for consuming messages from plugins,
-/// 
+///
 /// Can only be a single consumer per world,
-/// 
+///
 #[derive(SystemData)]
 pub struct Listener<'a>(
     Write<'a, Receiver<StatusUpdate>, EventRuntime>,
